Fixing renaming error with direct image comparison file
